src/main/java/com/dmwl/guacamole/net/encryptedurl/Security.java
package com.dmwl.guacamole.net.encryptedurl;

import javax.crypto.Cipher;
import javax.crypto.spec.SecretKeySpec;

import org.apache.commons.codec.binary.Base64;

public class Security {

	private SecretKeySpec secretKey;

	public Security(String secretKey) {
		this.secretKey = new SecretKeySpec(secretKey.getBytes(), "AES");
	}

	public String encrypt(String input) {
		byte[] crypted = null;
		try {
			Cipher cipher = Cipher.getInstance("AES/ECB/PKCS5Padding");
			cipher.init(Cipher.ENCRYPT_MODE, this.secretKey);
			crypted = cipher.doFinal(input.getBytes());
		} catch (Exception e) {
			System.out.println(e.toString());
		}
		return new String(Base64.encodeBase64(crypted));
	}

	public String decrypt(String input) {
		byte[] output = null;
		try {
			Cipher cipher = Cipher.getInstance("AES/ECB/PKCS5Padding");
			cipher.init(Cipher.DECRYPT_MODE, this.secretKey);
			output = cipher.doFinal(Base64.decodeBase64(input));
		} catch (Exception e) {
			System.out.println(e.toString());
		}
		return new String(output);
	}
}
